+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - Symlink for `pyvela/examples` in the root directory
 - Instructions for updating `Vela.jl` and `pyvela`
 - arXiv link in README
-- CITATION file 
+- CITATION file
+- In `SPNTA.from_pint()`, compute planetary ephemerides if they are absent in the input `TOAs` object.
 ## Changed
 - Residual plot in `pyvela-plot` script
 ## Fixed